+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [Unreleased]
 
-- Add `partial` expector
+- Add `partially` expector
 
 ## [0.1.0] - 2025-09-19